+++ lend/src/components/ActionMMModal/index.tsx
@@ -3,19 +3,16 @@ import { Row } from "react-bootstrap";
 import colors from "shared/lib/designSystem/colors";
 import styled from "styled-components";
 import { components } from "../../designSystem/components";
-import { Title } from "../../designSystem";
 import sizes from "../../designSystem/sizes";
 import { VaultOptions } from "../../constants/constants";
 import { CloseIcon } from "shared/lib/assets/icons/icons";
-import { fadeIn } from "shared/lib/designSystem/keyframes";
-import { css } from "styled-components";
 import Header from "./Header";
 import Footer from "./Footer";
 import Hero from "./Hero";
 
 const borderStyle = `1px solid ${colors.primaryText}1F`;
 
-export const FixedContainer = styled.div`
+const FixedContainer = styled.div`
   position: fixed;
   display: flex;
   justify-content: center;
@@ -43,77 +40,7 @@ const HeroContainer = styled.div`
   }
 `;
 
-export const HeaderRow = styled(Row)`
-  height: ${components.header}px;
-  border-bottom: 1px solid ${colors.border};
-  z-index: 1;
-  margin-left: 0px;
-
-  > * {
-    padding: 0;
-
-    &:not(:last-child) {
-      border-right: 1px solid ${colors.border};
-    }
-  }
-`;
-
-export const FooterRow = styled(Row)`
-  min-height: ${components.footer}px;
-  border-top: 1px solid ${colors.border};
-  box-sizing: content-box;
-`;
-
-export const HeaderContainer = styled.div`
-  display: flex;
-  margin: auto;
-  width: 100%;
-  justify-content: center;
-  border-radius: 0;
-  min-height: ${components.header}px;
-  border-bottom: 1px solid ${colors.border};
-  z-index: 1;
-  color: ${colors.primaryText};
-
-  > * {
-    margin: auto 0;
-
-    &:not(:last-child) {
-      margin-right: 8px;
-    }
-  }
-`;
-
-export const HeaderText = styled(Title)`
-  font-size: 16px;
-  line-height: 20px;
-`;
-
-export const DetailContainer = styled.div<{ show?: boolean; delay?: number }>`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  text-align: center;
-  height: 100%;
-  ${({ show, delay }) => {
-    return (
-      show &&
-      css`
-        opacity: 0;
-
-        &:disabled {
-          opacity: 0;
-        }
-
-        animation: ${fadeIn} 1s ease-in-out forwards;
-        animation-delay: ${delay || 0}s;
-      `
-    );
-  }}
-`;
-
-export const Content = styled(Row)`
+const Content = styled(Row)`
   height: calc(100% - ${components.header}px - ${components.footer}px);
 
   @media (max-width: ${sizes.lg}px) {
